+++ blocks/CloudImageEditor/src/EditorButtonControl.js
@@ -9,6 +9,7 @@ export class EditorButtonControl extends Block {
     title: '',
     icon: '',
     'on.click': null,
+    'title-prop': '',
   };
 
   initCallback() {
@@ -38,7 +39,7 @@ export class EditorButtonControl extends Block {
 }
 
 EditorButtonControl.template = html`
-  <button type="button" role="option">
+  <button type="button" role="option" l10n="@title:title-prop">
     <uc-icon set="@name: icon;"></uc-icon>
     <div class="uc-title" ref="title-el">{{title}}</div>
   </button>