+++ blocks/UploadList/UploadList.js
@@ -200,8 +200,13 @@ export class UploadList extends UploaderBlock {
 
 UploadList.template = html`
   <uc-activity-header>
-    <span class="uc-header-text">{{headerText}}</span>
-    <button type="button" class="uc-mini-btn uc-close-btn" set="onclick: *closeModal">
+    <span aria-live="polite" class="uc-header-text">{{headerText}}</span>
+    <button
+      type="button"
+      class="uc-mini-btn uc-close-btn"
+      set="onclick: *closeModal"
+      l10n="@title:a11y-activity-header-button-close"
+    >
       <uc-icon name="close"></uc-icon>
     </button>
   </uc-activity-header>